src/migrations/sheet2sheet_migrate/pyproject.toml
[project]
name = "sheet2sheet-migrate"
version = "1.0.0"
description = "Sheet-to-Sheet migration tool for Archive Metadata -> Iconik_Full_Metadata"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
